Big changes
Added monospace style.
Added an html info text.
Changed the style of the empty squares from X to [].
Added 2 arrays. 1 for unmoved pawns and 1 for legal moves.
Changed how the board is being printedwith a loop.
Added different turn-functions for white and black.
Created a while-loop to make the moves and check if the move is legal.
Started on functions for the pieces.